staticanalysis/bot: Merge RemoteWorkflow in Workflow, fixes #2081
