--- src/staticanalysis/bot/static_analysis_bot/workflows/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-from static_analysis_bot.workflows.base import Workflow  # noqa
